(search) Use same store configuration script for dev and prod mode and use REDUX_LOG environment variable
