+++ applications/search/Dockerfile
@@ -21,7 +21,8 @@ RUN npm run build
 
 #RUN npm test
 
-#ENV NODE_ENV production; just for local test purpose, variable set by OpenShift
+# use for local test purpose, variable set by OpenShift
+# ENV NODE_ENV production
 
 EXPOSE 3000
