+++ applications/search/src/config.js
@@ -1,0 +1,3 @@
+export const config = {
+  reduxLog: process.env.REDUX_LOG==="1"
+};
