+++ applications/search/src/index.jsx
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { BrowserRouter, Route } from 'react-router-dom';
 import ReactGA from 'react-ga';
 
-import configureStore from './redux/store/configureStore';
+import { configureStore } from './redux/configureStore';
 import App from './app/app';
 
 if (window.location.hostname.indexOf('fellesdatakatalog.brreg.no') !== -1) {
