--- applications/search/src/redux/store/configureStore.prod.js
@@ -1,10 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import api from '../middleware/api';
-import rootReducer from '../reducers/index';
-
-const createStoreWithMiddlewares = applyMiddleware(thunk, api)(createStore);
-
-export default function configureStore(initialState) {
-  return createStoreWithMiddlewares(rootReducer, initialState);
-}
