+++ applications/search/webpack.prod.config.js
@@ -56,7 +56,8 @@ module.exports = {
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
-        NODE_ENV: JSON.stringify('production')
+        NODE_ENV: JSON.stringify('production'),
+        REDUX_LOG: JSON.stringify(process.env.REDUX_LOG)
       }
     }),
     new webpack.optimize.UglifyJsPlugin({
